refactor indexer status components to match design
