+++ src/components/Status/IndexerStatusIcon.module.scss
@@ -1,0 +1,10 @@
+.low {
+  color: var(--color-success)
+}
+.medium {
+  color: var(--color-warning)
+}
+.high {
+  color: var(--color-error)
+}
+
