+++ src/components/Status/IndexerStatusLabel.module.scss
@@ -1,0 +1,10 @@
+.root {
+  :first-child {
+    margin-right: 10px;
+  }
+}
+.loader {
+  display: inline-block;
+  vertical-align: sub;
+}
+
