+++ src/types/IndexerStatus.ts
@@ -1,3 +1,5 @@
+export type IndexerStatusSeverity = "low" | "medium" | "high"
+
 export interface IndexerStatus {
   level: number
   id: number
